move `insertListOrTaskItemAt` logic into a reusable utility function
